Fixed a bug of overflowing processedBlocks

+++ src/liblightmetrica/lighttrace.cpp
@@ -101,7 +101,7 @@ bool LighttraceRenderer::Impl::Configure( const ConfigNode& node, const Assets& 
 bool LighttraceRenderer::Impl::Render( const Scene& scene )
 {
 	auto* masterFilm = scene.MainCamera()->GetFilm();
-	std::atomic<int> processedBlocks(0);
+	std::atomic<long long> processedBlocks(0);
 
 	signal_ReportProgress(0, false);
